resources small, make services file
